+++ src/main/java/performance/model/AnalyzeResult.java
@@ -3,9 +3,9 @@ package performance.model;
 public class AnalyzeResult {
     private final String analyzeName;
     private final long durationNano;
-    private final long iterations;
+    private final int iterations;
 
-    public AnalyzeResult(String analyzeName, long durationNano, long iterations) {
+    public AnalyzeResult(String analyzeName, long durationNano, int iterations) {
         this.analyzeName = analyzeName;
         this.durationNano = durationNano;
         this.iterations = iterations;
@@ -19,7 +19,7 @@ public class AnalyzeResult {
         return durationNano;
     }
 
-    public long getIterations() {
+    public int getIterations() {
         return iterations;
     }
